+++ src/components/ChartLabel.tsx
@@ -13,61 +13,46 @@ export interface ChartLabelProps
         Animated.AnimatedProps<ViewProps> {
     alignX?: Alignment2D['x'];
     alignY?: Alignment2D['y'];
-    textWidth?: number;
-    textHeight?: number;
-    ignoreBounds?: boolean;
 }
 
 const ChartLabel = (props: ChartLabelProps) => {
     const {
         alignX = 'center',
         alignY = 'center',
-        textWidth,
-        textHeight,
         title,
-        numberOfLines,
         textStyle,
         style,
         render,
         ...otherProps
     } = props;
 
-    const isFixedWidth = (textWidth || 0) > 0;
-    const isFixedHeight = (textHeight || 0) > 0;
-
-    const { ignoreBounds = isFixedWidth || isFixedHeight } = props;
-
     const textProps: Animated.AnimatedProps<TextProps> = {
         selectable: false,
         style: [{ textAlign: alignX }, textStyle],
-        numberOfLines,
     };
+
+    let content: React.ReactNode;
+    if (render) {
+        content = render(textProps);
+        if (typeof content === 'string') {
+            throw new Error(
+                'Use the title prop to display a basic text string'
+            );
+        }
+    } else {
+        content = <Animated.Text {...textProps}>{title || ''}</Animated.Text>;
+    }
 
     return (
         <Animated.View
             {...otherProps}
             style={[
                 styles.container,
-                {
-                    alignItems: kAlignItemsMapX[alignX],
-                    justifyContent: kAlignContentMapY[alignY],
-                },
+                { justifyContent: kAlignContentMapY[alignY] },
                 style,
             ]}
         >
-            <Animated.View
-                style={[
-                    ignoreBounds ? { alignSelf: kAlignSelfMapX[alignX] } : {},
-                    isFixedWidth ? { width: textWidth } : {},
-                    isFixedHeight ? { height: textHeight } : {},
-                ]}
-            >
-                {render ? (
-                    render(textProps)
-                ) : (
-                    <Animated.Text {...textProps}>{title || ''}</Animated.Text>
-                )}
-            </Animated.View>
+            <Animated.View>{content}</Animated.View>
         </Animated.View>
     );
 };